+++ src/header/index.js
@@ -1,4 +1,2 @@
 export Basic from './header-basic.component';
 export Custom from './header-custom.component';
-export Navigation from './header-navigation.component';
-export TabNavigation from './header-tab-navigation.component';